cut a segment defined by a range in 1D coordinate into n pieces

cut a segment defined by a range in 1D coordinate into n pieces

segment_range(range = c(0, 1), n, scale = NA, return.cutter = T)

Arguments

range
a vector of 2 numeric values defining a segment

n
cut the segment into n pieces

scale
a vector of n numeric values to control segment sizes relatively. Set NA or NULL for equally dividing.

return.cutter
only return the coordinate values of internal cutters

Value

a vector of n+1 or n-1 if return.cutter

Examples

segment_range(c(0,10), 3, scale = c(1, 2, 7), return.cutter = F)

#> Error in segment_range(c(0, 10), 3, scale = c(1, 2, 7), return.cutter = F): 没有"segment_range"这个函数