Transformation by a given function

Transformation by a given function

tf_fun(
  object,
  fun,
  along = c("x", "y", "xy"),
  layout = "transforming",
  xrange.from = NULL,
  yrange.from = NULL,
  xrange.to = NULL,
  yrange.to = NULL,
  crosses = NULL,
  ...
)

Arguments

object
a CrossLink object

fun
function for coordinate transformation

along
another axis will changed along this given axis. Such as, if along = "x", transformation function will change y values along x axis and keep x values asis.

layout
layout name to save

xrange.from, yrange.from
the range to be scaled from before transformation by function

xrange.to, yrange.to
the range to be scaled to

crosses
only these crosses will be transformed

...
other parameters passed to fun

Value

an updated CrossLink object

Examples